--- src/components/TonConnectButton/index.tsx
@@ -1,8 +1,0 @@
-'use client';
-import { TonConnectButton as BaseTonConnectButton } from '@tonconnect/ui-react';
-
-export function TonConnectButton() {
-  return (
-    <BaseTonConnectButton />
-  );
-}
